refactor(manager): consolidate the singular service/ module into services/
Remove the legacy singular `service/` package and move its contents under
the canonical `services/` package:

- service/base.py -> services/base.py (ServicesContext)
- service/container_registry/harbor.py -> services/container_registry/harbor.py

Update all `ai.backend.manager.service.*` imports to `services.*` and
convert the moved base module to an absolute import.
