+++ src/ai/backend/manager/services/container_registry/harbor.py
